sndtrack:  Eliminate clicks on new four-tone pitch
The relationship between the number of samples that have elapsed and the
position of the phase marker is dependent on the playback rate -- if the
pitch (and therefore, the rate) changes, the computed phase marker jumps
some distance.  The solution is to maintain the phase markers instead of
recomputing them.