vlib:  Meld `else`, `if` into single `else if` op
Both operands to `else` are supposed to be blocks (or at least something
invocable), but an exception is made for an `if` expression as the right
operand, which is automatically upgraded to a block expression.  Instead
of deferring the repair of internal consistency until execute(), combine
`else if` into a single op node in Analyzer::visit().

This makes it possible to fold `if` expressions.  Previously, we weren't
able to do so, since `if x then {abcd} else if y then {defg}` would fold
to `if x then {abcd} else ()` assuming `y` is known to be false, whereas
we'd need `if x then {abcd} else {}` to make the right operand of `else`
a block, and the folding code working on an `if` node wouldn't know what
the parent op was so it couldn't discriminate.  But if `else if` ops are
used instead, then folding can proceed.