vlib:  Make `and` and `or` right-associative
Semantically, this is a no-op; the difference is in run-time complexity.

The expression `a and b and c` previously parsed as `((a and b) and c)`.
Suppose `a` is false.  That short-circuits the entire expression, but it
still requires descending all the way down the tree to arrive at the `a`
node before evaluating it.  Whereas with `(a and (b and c))` we can look
at `a` immediately.