ams-ui:  Manage the editField in SetDItem()
If the editField ceases to be an editText, shrink its text handle to fit
and clear editField to -1.  (A TERec has a field to store the length out
of band, but a mere dialog item has just the handle.)

If we're setting the type to editText and there's no editField currently
set, set it.  In addition, if no TextEdit record exists yet, create one.