ams-ui:  Check for null editField on activate
There are two bugs fixed here.  One, if the only editItem has changed to
a statItem, then textH will exist but editField will be -1, and the call
to get_editField() will return NULL -- so reading the `bounds` member is
problematic.  Two, the TextEdit record would become active, allowing the
caret to blink even though there's no editable field present (if we only
guarded the assignments and not the TEActivate() call, that is).