Genie:  Install separate 68K exception handlers
The 68000's exception stack frames don't provide the vector offset.  :-(

Instead of pulling the vector offset out of the stack frame, now we have
to install unique exception handlers that already know which vector they
belong to.  Since the vector offset is known at compile time, we replace
the `(A0,D0.W)` read with a plain `(A0)`.  However, we still have to set
D0 to the offset, since generic_68k_recovery_handler() still performs an
indexed memory read of the user handler table.

I would have preferred to write an asm function template, but this seems
to be beyond the abilities of MWC68K to comprehend.  Therefore, we apply
brute force and use macros.