Pedestal:  Don't direct-dispatch an AE on Quit
This is needless bloat that only affects recording events for the Script
Editor.  Other than Quit itself, AE recording has never been implemented
and therefore falls under YAGNI (You Aren't Gonna Need It).  It would be
a non-trivial amount of work for uncertain benefit, and it would even be
harmful in RAM-constrained environments.