poseven:  Fix g++ 4.0.1 `struct stat` regression
Because `stat` names a function as well as a type, we're forced to write
the `struct` keyword when referring to the latter.  However, Apple's g++
4.0.1 can't handle a function template returning `struct stat` -- though
it does work if we write `struct ::stat` instead.  But that wreaks havoc
for Apple's g++ 3.3.  So instead, we have a compromise that everyone can
cope with, the `struct_stat` typedef.