ams-core:  Honor Shift and Alpha modes case-wise
This isn't a complete solution, but it's enough for entering a name into
Lode Runner's high scores dialog.

Note that the Linux front end currently understands the Alpha mode to be
in effect only while the key is down, like any other modifier key.  This
would work for a physically locking Caps Lock key (or perhaps a keyboard
that mimicked one signal-wise), but needs to be changed for the majority
case.  MacRelix already sends upper-case letters, so this doesn't affect
it -- though maybe we should reconsider and have it send only unmodified
characters, to match the other front ends.