vx:  Garbage-collect threads before calling exit()
We can't rely on static destruction, due to the arbitrary order in which
groups of static destructors are called.  Everything works as long as we
call cull_unreachable_objects() before destroying joinable_threads_mutex
from thread_state.cc.  Since the static destruction of tracker.cc (which
calls cull_unreachable_objects() one last time) can't be depended on (as
it might happen too late), we have to call cull_unreachable_objects() an
additional time from main(), to ensure that the above-mentioned mutex is
still valid when the thread objects get destroyed.  (The vector that the
mutex guards will have already been emptied, but the mutex must be valid
nevertheless, so it can be locked.)

Mac OS X 10.3's g++ 3.3 is the first toolchain encountered which exposed
this issue.