portaudio-pkg.vx:  Only make lib/libportaudio.la
By default, `make` makes `all`, which includes tests and examples.  This
hasn't been a problem in stock OS X configurations, but a MacPorts-based
setup failed to compile one of the test programs, breaking our own `make
sndtrack`.  However, all we need is the static library libportaudio.a --
there's no target for that explicitly, but if we target libportaudio.la,
libportaudio.a also gets built.  This fix also saves a little build time
on OS X installations without this problem.