[NEW] Amethyst:  Fullscreen CG-based AMS front end
Amethyst is a Quartz-based application and Advanced Mac Substitute front
end project.  It achieves 64-bit compatibility by forgoing windowing and
instead relying on Core Graphics display capturing.  Unlike the existing
front ends, which have to be instructed at run time which application to
launch, Amethyst is designed as a template, to be combined with a single
68K application into a self-contained, pre-configured, runnable package.