[NEW] retromatic.vx:  Per-app emulator generator
One of Metamage Labs' recent creations is the Retro-Matic App Dispenser.
The way it works is very interesting:  When the program is executed from
the command line (with the output directory and an application specified
as arguments), it delivers a new Mac OS X application, whose function is
to provide a Mac emulation that's almost, but not quite, entirely unlike
Cocoa.  To do this, it actually runs Advanced Mac Substitute -- twice --
to extract the application's creator code (from its 'BNDL' resource) and
its icon, which (along with other, optional configuration data) are then
incorporated into the new application's bundle structure.  The result is
individually tailored to meet the 68K application's precise requirements
for emulation and comfort.  The practical upshot of all this is that you
can double-click the icon of a classic Mac application and instantly run
it on any Mac released this millennium, running any version of Mac OS X.

Share and Enjoy!