sndtrack:  Fix distortion in square-wave tones
The previous code was generating square waves that cut off at the end of
a buffer instead of continuing in the next buffer.  The exact details of
how the sign of the next run of samples was determined changed recently,
but in any case, a buffer ending with a short run was always followed by
a full-length run instead of the complementary short run, causing noise.