ams-qd:  Relax region validity tests so TFE passes
One of the puzzles in The Fool's Errand (The Singer) has flawed regions,
but will still work if we allow them.  In two cases, the bottom point of
the region is one pixel above the bbox.bottom.  In the other, the top of
the region is one pixel below bbox.top.  Note that it's yet undetermined
whether The Fool's Errand or AMS is to blame for the faulty region data.