WDEF-1:  Don't unclip to GrayRgn
Not only would window frames draw over the menu bar and rounded corners,
but moving the window to or from a location partly offscreen would force
the entire window to be redrawn (even if no new area was exposed).  This
also made updates in Mines erase entire windows but repaint selectively.