ams-qd:  Always lock screen updates in redraw_lock
The raster lock within redraw_lock (which locks the screen and hides the
cursor) is conditional on the cursor being in the work area.  Otherwise,
the raster lock is inactive and has no effect.  Indeed, there's no point
erasing and repainting the cursor in that case, but we should still lock
the screen so that just a single update happens after drawing completes,
if the redraw_lock's destination bitmap happens to be the screen buffer.