ams-snd:  Allow longer 22kHz sounds by looping
If a sampled sound with a 22 kHz sample rate exceeds 30000 samples, chop
it into multiple parts (each part having a multiple of 370 samples (with
the possible exception of the last one)), and send each one to the Sound
Driver in a separate call.  Now we can play the 'kaboom' sound in Mines.