ams-ui:  Trim editText handle in GetDItem()
Generally, we don't keep TextEdit text handles exactly sized to fit, and
instead we rely on the teLength field.  But when the caller only has the
handle to go on, it has to be appropriately sized.  Though we could have
instead called recover_dialog() to retrieve the text size in GetIText(),
that's a lot of effort just to eliminate one SetHandleSize() call, which
doesn't require reallocation anyway.