ams-qd:  Include line-based polygons in regions
Inside Macintosh, Volume I is really vague on how this is meant to work.
Though it does mention that defining a region and defining a polygon are
mutually exclusive -- which means we can safely use the polygon-defining
machinery ourselves for defining a region.  We're going to disregard any
contacts between line segments and framed shapes, and simply collect all
line segments into a polygon, which we then merge into the final region.