ams-qd:  Clip CopyBits to source bitmap's bounds
System's Twilight's default animation mode copies pixels from the screen
into a work buffer, but it doesn't explicitly clip to the screen bitmap,
instead relying on either the clipping to occur implicitly or the memory
above and below the screen to be readable (harmlessly filling those rows
of the work buffer with garbage that will never be displayed).  Since we
tend to strictly enforce memory access, implement the expected clipping.

Otherwise, we blow up when Mazer jumps offscreen.  And I do mean "bang".