vlib:  "Spoil" arrays when assigning new elements
If we don't do this, we can wind up with an array flagged as cycle-free,
even after assigning to it an object that's not.  The next time GC runs,
the root-marking phase will skip over the array (flagged cycle-free) and
everything in it, leaving the newly stored object marked for destruction
even though it can still be reached.  The rest, as they say, is history.

The second spoil() call in targets.cc is sufficient to pass the first of
the test cases; both calls are necessary to pass the second test.  There
may be other variants of this error lurking, but these are the ones with
formerly failing tests now passing, and no others are currently pending.