varyx-posix:  Always trace through thread results
Though there won't be anything stored in a thread result field until the
thread terminates, tracking the symbol without tracing it during marking
causes it to expire, which in turn causes it /not/ to expire later, thus
resulting in a memory leak.  While we could move the track_symbol() call
from the thread_state constructor to pthread_start(), a simpler and less
invasive fix is to just remove the terminated() condition, which has the
added (or subtracted?) benefit of eliminating a mutex locking operation.

This fixes the memory leak that 357ed94f8405fbe11429f303a6572b78e4a14a7c
exposed in v/vx/t/thread-reference-cycle.t when running within valgrind.