ams-core:  Dequeue events properly
We use qType as a mutex, but we were only clearing it in GetOSEvent() --
any other dequeueing, viz. PostEvent() (when the event queue is full) or
FlushEvents() (the main source of trouble) would leave elements unusable
for new events, until find_empty_slot() couldn't and tried to discard an
element from the head of what was actually an empty queue, which is Bad.