ams-ui:  Don't ParamText-reexpand editText items
Reexpansion of editText items makes Lemmings draw garbage (starting with
the four right guillemots marking the end of a handle's allocated space)
in its access code entry field after submitting a valid code.  According
to Inside Macintosh we shouldn't be initially expanding them either, but
that's a more invasive change and not necessary to fix the Lemmings bug.