ams-core:  Fix the KeyMaps bitfield mappings
For an intended virtual keycode K, we were instead getting (K ^ 7) -- so
the byte-swapping produced by subtraction was correct, whereas reversing
the bits wasn't.  Declare KeyMaps as an array of bytes instead of longs,
which not only makes this byte-swapping unnecessary, but /also/ requires
fewer instructions (even before removing the faulty bit-reversal logic).