ams-core:  Avoid double-free in DisposePtr
The sound/music engine in Lemmings double-frees the vblAddr jump island.
Without this check in place, the second call to DisposePtr calls through
to MSL's free() and corrupts the free list.  (Possibly all it does is to
add a second instance of the block to the free list, permitting malloc()
to return it twice -- which is the problem witnessed in Lemmings' case.)