quickdraw:  Improve 68K codegen by eliding shifts
Subtraction of pointers to word-sized objects implicitly divides by two,
but the generated code implements a /signed/ division -- even though the
dividend is assured to be non-negative in context -- which although more
efficient than a DIVS instruction, is also more verbose -- consisting of
half a dozen instructions (five of them bitshifts).  But in fact we have
no real need to right-shift the pointer difference, since we can instead
left-shift the bit mask at O(0) cost.