ams-fs:  Fix macbinaryfs GetFileInfo catastrophe
What happens depends on the value previously stored in fdFlags:  A value
of $0100 causes an immediate crash (accessing unmapped address $000101);
a value of $2100 clobbers _SetItmIcon's entry in the Toolbox trap table.

If the byte at offset 74 hadn't been defined as a zero byte in lieu of a
proper format-identifying magic number, this bug wouldn't have happened.