ams-ui:  Fix ParamText() in invisible dialogs
Since 4fe9e47a103b11b45eda68dc8b0a1fc68366182a, text in GunShy's initial
showing of the About box (on launch) doesn't display.  The basic problem
is that ParamText() has to consult the WindowList because it isn't given
a reference to the dialog, and it was dandy until we began filtering out
invisible windows (like GunShy's dialog box when ParamText() is called).