v68k-alloc:  Fix multi-page allocation bug
The allocator depends on monotonically increasing addresses to recognize
the bounds of an allocation, so in addition to locating n contiguous and
available pages, we also have to ensure that the neighboring pages don't
contain the corresponding next and previous values, lest the deallocator
confuse two unrelated allocations that happen to have continuous address
ranges for a single allocation and mark as deallocated pages that are in
fact still in use.  Our post-bounds continuity check was producing a few
false negatives for multi-page allocations.  As always, hilarity ensued.

Specifically, applying 361cf6e2ef2494c9583eb8c0f434d2fb194b06d6 triggers
a crash on launch in Prince of Persia more often than not, prior to now.