ams-qd:  Honor non-zero firstChar of FontRec
This fixes the rendering of fonts whose FontRec's firstChar is non-zero,
including MacPaint's font for the tools palette.  The last tool actually
doesn't get drawn at all, but this is because pnLoc.h is somehow off for
just that one case:  It should be alternating between 21 and 47, but the
last tool is drawn at pnLoc.h = 83 for some unknown reason.  If the code
is hacked to fudge it back to 47, the FillPoly icon renders as expected.