Amethyst:  Await Open Documents/Application event
Wait for either an Open Application or Open Documents Apple event before
capturing the screen and entering the main event loop.  (This means that
using AppleScript's `launch` command will launch the application process
itself, but not the emulator (until it receives one of the two events).)

If an Open Documents event (received either on launch or later) precedes
any Open Application event, then that information will be passed through
the AMS_OPEN environment variable and added to AppParmHandle by ams-seg.