vlib:  Store MB32's uint32_t data as native endian
This change fixes ordering on little endian architectures, avoiding byte
swaps in the comparator at the cost of adding them for I/O.  To be fair,
we haven't done any mb32 ordering yet, but it's not out of the question,
and anyway the performance impact is likely to be negligible either way.
In the meantime, this brings MB32 in line with the way Mac OS X natively
stores OSTypes, which will be convenient when implementing certain APIs.