A-line:  End cross-built cpp filenames with .i/.ii
That is, use .i and .ii for preprocessed translation units only if we're
running A-line outside of MacRelix (and we only preprocess separately if
we're targeting classic Mac OS, which means we're cross-building in this
case).  Continue using .cpp when running in MacRelix, since .i/.ii would
confuse the Metrowerks c++ compiler as unrecognized filename extensions.
(The cross-builder strips .i/.ii as it copies files, leaving .c or .cc.)