vlib:  Replace infix `(+)` with `~` as bitwise XOR
The existing operator token is supposed to resemble the mathematical XOR
symbol `⊕` (which admittedly it somewhat does), but it's not immediately
obvious what it means, and the asymmetry with `&` and `|` with regard to
token length is offputting.  Plus, `(+)` is a nuisance to type.  The `~`
token, on the other hand, relates to its unary lookalike in much the way
that subtraction does to additive inverse (both `-`), and even its usual
position on US QWERTY keyboards places it above a glyph that's typically
rendered as a diagonal line dexter, just like the non-exclusive OR, `|`.