sndtrack:  Switch to 16-bit output samples
This is not a no-op.  This more than doubles the output resolution of an
individual four-tone voice, from 6 bits to 14 bits (a 133% improvement).

Square-wave and free-form sounds should be indistinguishable from before
this change, on the other hand.  Although a future change to mix samples
from free-form waves would retroactively also benefit from the increase.