sndtrack:  Diminish last four-tone output buffer
The synth handler function result is no longer a sample count.  Negative
values still mean that the input was already exhausted and that the next
sound node should be played instead.  Positive values also have the same
behavior.  What's different is that a zero result now indicates that the
a fadeout should occur, which is important for four-tone sounds in order
to avoid a popping noise when they reach the end of their allotted time.