vlib:  Disallow `a? b` for prefix unary operators
We have to be more careful here than with postfix operators, because the
prefix operators do stack -- e.g. `**p` for a double dereference, `*++p`
for dereferenced preincrement, and other possibilities.  An exception is
the key/var macro operator, which can only be preceded by the list-count
operator (`(+)`), which typically carries intervening whitespace anyway.
(Note that there's no reason to even do this, since it just yields `1`.)