Amethyst:  Factor out cursor warp, fix edge moving
Previously, when the mouse cursor hit the screen's right or bottom edge,
it stopped moving, as if halted by excessive friction.  As it turns out,
it's a known Core Graphics issue with several suggested remedies.  We've
picked CGAssociateMouseAndMouseCursorPosition(), which works as expected
in macOS 10.12, at least.  In Mac OS X 10.5, it causes the first warp to
fail if the mouse is moving during display capture, so we skip that one.