vlib:  Avoid overspoiling with make_target() calls
There are two different things happening here.  One, we have to find the
actual Value slot that we're assigning to, e.g. the nth element of array
x.  Two, the objects above that one in the containment hierarchy must be
potentially marked non-cycle-free.  The problem is that `<->` (swap) has
two targetable operands, and whether to spoil one of them depends on the
value stored in the other -- a cyclic data dependency.  In that case, it
may be necessary to call make_target() twice for the left operand, since
the first call is made before we know what value should be assigned, and
hence whether spoiling is called for.  If so, a second call is required.

While `<-` (move) doesn't have the above cyclic data dependency, it does
have dual lvalue operands, and whether the left operand's parents should
be spoiled depends on the value of the right operand, not the expression
which references it.  This change is observed by the included test case.