ams-rsrc:  Don't release 'INIT' resource afterward
Disposing the resource handle tends to interact poorly with trap patches
(which require code to stay resident).  So far, it breaks Solitaire, but
only when launched from Solitaire/rsrc, not Solitaire.mbin.  Some others
work just fine.  It would be reasonable to release a resource only if it
hasn't been detached, but not releasing at all is simpler and works now.