vxo:  Add StrPool type for interning strings
A StrPool is a repository for interned strings.  An input string will be
copied into one of several "lane" buffers, which don't move in memory --
allowing the string's address to be returned and used thereafter for the
life of the program.  StrPool is not a subtype of Box and isn't designed
to be contained or copied; usually there should be only one per program.