libpthread:  Don't store tid before thread created
When pthread_self() is called and no secondary threads have been created
(as happens when single-threaded programs call raise()), don't store the
tid for the thread leader, as doing so will cause a problem with fork():
The stored tid won't match the newly forked child process.  This remains
an issue in multithreaded programs that call fork(), but these aren't as
common, and the partial fix is worth doing.