ams-ui:  Fix Standard File package call prototypes
Any implemented Pack3 call, even if cancelled, would overwrite a word on
the stack just beyond the arguments.  In Lode Runner, this is the return
address at which execution resumes after the File->New-handling routine.

This fixes Standard File invocations for Lode Runner, The Fool's Errand,
and Prince of Persia, at minimum (and probably for some others as well).