skif:  Load the CLUT with a default gray palette
Note that this table runs from black to white -- /unlike/ classic Mac OS
grayscale, which is white to black.  Be warned that this may yet change:
The intent here is to provide an arbitrary non-repeating set of entries,
improving on the previous all-black set -- not a known set of constants.