interact-x11:  Fix big-endian/mixed reversed image
By specifying the bit and byte orders in the XImage struct, we avoid the
need to reverse each group of eight pixels manually.  This in turn means
we can drop the pixel buffer entirely and pass the shared buffer address
directly to XCreateImage(), provided that we null out the field prior to
calling XDestroyImage() (which otherwise tries to free() it, which makes
glibc unhappy).  This fixes the display not only on big-endian machines,
but also between systems with different endianness using X11 forwarding.

Modified-By:  Josh Juran <jjuran@metamage.com>
Approved-By:  Andrew Rogers <tuxlovesyou@gmail.com>