retromatic.vx:  Choose Amethyst for arm64 hosts
Amphitheatre is based on OpenGL and uses considerably less CPU time than
Amethyst, which is based on Core Graphics (a.k.a. Quartz) and QuickDraw.

However, we have a report that Amphitheatre fails to work in macOS 11 on
ARM, whereas Amethyst does work.  Nonetheless, we've also confirmed that
Amphitheatre works in Big Sur on x86 -- so continue to use it there, and
substitute Amethyst on ARM.