Genie:  Prefer mempcpy() to memcpy()
Calling mempcpy() is smaller because it doesn't get inlined.  As for run
time, we avoid the branch logic of memcpy() deciding whether to make the
trap-dispatched call to BlockMoveData() (which it won't for SPIEL's tiny
messages).  For v68k emulation, we could optimize a simple function call
more easily than an inlined call site.