mac-sys-utils:  Improve ticks->us accuracy, speed
Instead of the conventional Macintosh frame rate of 60.15 (documented in
Inside Macintosh), use the considerably more accurate figure of 60.14742
for converting ticks to microseconds.  The other one loses 3.7s per day.

Avoid a 64-bit multiplication and division, using a single 68K-optimized
long multiplication and bit shifts.  This generates smaller object code,
and it could be faster if clock() or gettimeofday() is called frequently
in MacRelix running in System 6 or earlier (with no _Microseconds trap).