ams-ui:  Rename string_list draw/click variables
This no-op patch introduces a distinction between the 'selection' (which
is an index into an abstract list of strings) and a 'click index' (which
is the index into the set of list rows visible on screen, into which the
abstract string list is mapped and rendered).  This simplifies the next.