fndr-sync.vx:  Try FSRef conversion before FileURL
The former allows relative pathnames, which the latter does not.  We can
also pass it to the Finder as is instead of coercing it to another type.

This works in 10.5 (at least) and later, so all the 64-bit scenarios are
covered.  If coercion to FSRef fails, we proceed as before via furl/fss.