Revert "Tic-tac-toe:  Avoid QD erase damage to CG gridline"
This was the wrong fix.  The issue was that when the window's height was
an odd number of pixels, the vertical margin at the lesser Y value would
be one pixel less than the other vertical margin, and QuickDraw's notion
of which margin was margin.v pixels (top) conflicted with Core Graphics'
default of Y=0 at the bottom.  The correct fix was to flip the geometry.

This reverts commit 3e42d4c968215fb39e765abd0f2d651d9fbd685d.