ams-core:  Leave auto-vsync off by default
Auto-vsync exists so that Lode Runner's multiple blits per frame yield a
single screen update.  However, Tetris has animations consisting of many
blits, one right after another, which must each issue a distinct update.
It also installs a VBL handler for generating audio, which was enough to
run afoul of the heuristic for enabling auto-vsync.  So now it's opt-in.