ams-core:  Fix interrupt outage after VRemove fail
We were disabling interrupts but not reenabling them if VRemove() failed
to dequeue the task (because it wasn't queued).  This had been the cause
of audio playback gaps in Tetris, which attempts to VRemove the SoundVBL
as part of its sound setup.  Dequeue() is already serialized internally,
and while we do have to guard against concurrent access to VBLQueue, our
critical section doesn't in fact need to cover the prior Dequeue() call.