ams-pack:  Fix broken asm codegen of const int
So, let me see if I got this right:  In an asm context, `enum { n = 8 }`
means that `#n` evaluates to `#8`, but `const int n = 8` makes it `#0`??

So all this time we've been executing `CHK.W #0xFFFF,D0`.  But we didn't
notice because v68k was misevaluating the immediate operand by extending
it with zero bits instead of the sign bit.  Fixing that bug will convert
that instruction from a no-op to a fail-all, so we're fixing this first.