ams.vx:  Axe support for displayfs (FORGE) backend
Now that Advanced Mac Substitute runs real applications, it's simply not
practical to run it on the machines which are capable of running classic
Mac OS -- in MacRelix, at least.  It might make sense on a G4, with more
code running natively and a custom front end, but this approach -- which
to its credit was the one used to develop the initial Mac OS and Toolbox
support that allowed test applications to run -- has since then outlived
the benefit it provided.  FORGE is one of the coolest ideas of mine that
didn't quite work out well, but at least ams.vx showed what it could do.