NyanochromeCat:  Rewrite timer to be frame-unaware
The animation timer knows about time, but has no idea which frame is the
current or next one.  This change avoids a bug when rewinding past start
frame 0 (where the `%` operator's behavior is unspecified, and maybe not
the modulo operation we wanted), and makes the timer forgiving of delays
in calling pulse().  Previously after a delay, it would skip frames; now
it will steal the lost time from the new frame if it's less than a pulse
interval (to keep pace without slippage), and reset the pace with a full
frame duration otherwise.  This will be more friendly to older machines.