ams-fs:  Incrementally write to docfs, enable 64K+
Make it possible to write files longer than 64K.  The one implementation
that exists so far uses the Freemount protocol, sending a single message
that's limited in size.  For small files, it's fine to accumulate writes
into a buffer and also upload the whole file each time, but that doesn't
scale at all -- it's O(n^2) in file size.  This hadn't been an issue yet
because saved game files are tiny (especially Tic-tac-toe's), and even a
maximally compression-resistant MacPaint document maxes out at 52K.  But
expanding an SEA (self-extracting archive) or other file container could
create arbitrarily large files -- the smallest disk images are 400K, for
example, which is already too big -- requiring us to use a new approach.

To handle Write requests to the Documents volume, use incremental writes
instead of whole-file writes at the Freemount layer.  However, writes to
appfs are unchanged for now, because we depend on automatically creating
the file.  We also need to preserve whole-file writes for docfs in order
to implement SetEOF requests, until we have a direct means to accomplish
this in Freemount.