vxo:  Don't unshare() in Container::begin()
We need to rethink how this is going to work, because nothrow semantics,
copy-on-write, and regular types don't mix.  In many cases, what we have
is the sole reference, so copy-on-write isn't needed.  In the future, we
may need to preemptively call (a yet-to-be-written) unshare_nothrow() to
ensure we don't violate constraints, and violations could panic (like an
assertion) instead of throwing, surfacing errors without major overhead.