xv68k:  Prevent writes to the boot code / OS block
That code is installed by xv68k itself, so no emulated code has any need
to write to it.  This provides a defensive check against stack overruns.

In particular, instead of a single Varyx test failing in xv68k (and that
only in non-turbo mode), we now have four tests failing (all recursive).