ams-qd:  Fix diameter=1 circular region edge case
When the diameter is one, the `y0 < x` loop condition evaluates as false
and the loop is never entered.  The single-cell scratchpad is sufficient
for generate_circle() in this case, but since the loop didn't initialize
it, we need to do so outside of the loop (in advance, for diameter > 1).