ams-qd:  Fix missing edge bytes using notFoo modes
When using CopyBits() with notSrcCopy, notSrcOr, notSrcXor, or notSrcBic
(or drawing with any of their pattern counterparts), any bytes that were
partially within the targeted pixel range didn't get drawn (i.e. the two
endpoints if not byte-aligned), and if the leftmost byte were lost, then
the subsequent bytes were drawn displaced to their left by eight pixels.

This was causing graphics corruption in the Finale of The Fool's Errand,
and just recently in SKIFView when it displayed monochrome-light images.