ams-qd:  Distinguish between lineTo, unique points
A lineTo point is a non-starting point that adds an edge to the polygon.
The starting point is never a lineTo point; all other points always are.
A unique point is one that doesn't close the polygon; a point closes the
polygon if it's the last point and it matches the starting point.  (If a
point matches the one before it, it won't get added in the first place.)