ams-io:  Fix end-of-sound-time calculations
If you Quit Lode Runner mid-game without having earned a new high score,
It plays the death sound and then immediately exits.  Our timeval adding
logic failed to carry the one when the microseconds field overflowed, so
in that case we were attempting to schedule a wakeup for a time that had
already expired.  Since the death sound lasts about 0.8s, it was getting
prematurely cut off around four times out of five.  (This is now fixed.)