ams-io:  Rewrite KillIO logic
Don't call IODone(), which calls the driver to initiate further I/O, for
each request when it's at the head of the queue.  Instead, flush out the
queue all at once (atomically), and merely call the completion routines.

Atomically empty the queue all at once, and especially don't try to call
the