ams-ui:  Fix rampant highlighting from TEUpdate()
This fixes the regression from 851f9fd6688af4e5f3227cdcae3dd41af2c009dd,
which makes TEUpdate() update the selRect field.  Because of the way the
Dialog Manager shares a solitary TextEdit record among many text fields,
we have to temporarily mark the TE record inactive, except for the field
being edited.  It's also necessary to update the selRect when hiding the
selection, and since that means both callers of showhide_selection() are
now calling update_selRect(), that call can be factored out and moved to
showhide_selection() instead.