vlib:  Order multibyte fields natively by default
Don't require e.g. `i32` to be qualified as `native i32` to be used with
`vector`.  In the future, there will be a reason you might want to do so
anyway -- to lock in native byte ordering for one particular field while
others get qualified en masse as having a different ordering.  While not
quite an example of this exactly, struct sockaddr_in's sin_port field is
big-endian, while sin_addr.s_addr is natively ordered, so mixed-ordering
structs do exist.  In any case, the byte-order qualifiers (`big-endian`,
`little-endian`, and `native`) are not yet built into Varyx. which makes
it excessively inconvenient to use stdint types as vector element types.