xv68k:  Use fixed RAM as main graphics memory
This works around an issue with Vanlandingham:  It issues a write of 4K,
starting in but not limited to the page 2 screen buffer, which is beyond
what the screen memory translator will accept -- but only with fixed RAM
size configurations.  By shunting screen I/O to main memory, the dubious
write succeeds and Vanlandingham is happy.  (The subsequent update cycle
will fail, but the BlockMove callout only issues the update for its side
effects (if it has any), and doesn't care whether it succeeds or fails.)