ams-core:  Don't assume A1 is valid in Gestalt()
Apple's glue for Gestalt() invokes the _Gestalt trap with A1 pointing to
a location in which the response will be stored.  But our implementation
of the trap handler can't assume that it's been invoked by Apple's glue.

This fixes a probable code corruption issue in the Gestalt! application,
as indicated by the Illegal Instruction exception that no longer occurs.