ams-core:  Remove broken Button() polling throttle
This was unreasonably delaying legitimate ongoing calls to Button() used
in Spectre Challenger to poll the button state with some frequency.  The
poor behavior is calling Button() in a tight loop instead of waiting for
an event, and applications that do that can be patched to behave better.

(Spectre Challenger was essentially unplayable due to massively sluggish
turning of the player's tank.  This is now fixed.)