ams-qd:  Apply polarized palettes to PICT pixmaps
Rather than downsampling the palette indices, which may or may not yield
useful results (and, in fact, would even produce an entirely white image
if the palette has at most 128 entries), generate a polarized version of
the palette (whose entries are each either black or white) and transform
the image by replacing each byte with the palette entry corresponding to
that index.  This fixes Spectre's Vector/Filled button and About screen.