Pedestal:  Explicitly call {Begin,End}Update()
We're depending on window_update() not to throw, which would only happen
if view->Draw() did.  I'm not aware of any Draw() implementations having
such behavior.  MacRelix has some that subscript a std::map, which would
theoretically throw if allocating space for a new key failed, but in all
cases the key should be present before the View subclass is constructed.