Amethyst:  Define dummy overlay_enabled flag
We're not going to implement a screen overlay for Amethyst.  Amethyst is
a workaround for Amphitheatre not working when built for ARM.  All other
macOS front ends use OpenGL and can share a single implementation of the
screen overlay.  Amethyst would require not only a Quartz implementation
but a compositor layer (since we capture the display directly and bypass
the window manager).  For now at least, Amphitheatre works in Rosetta 2,
and that's the recommended solution while no better option is available.