aglow:  Prepare for next patch
Move shifted_max_zoom to global scope, so new functions can see it.  Let
shifted_current_zoom track the current zoom level, in a compressed form.
Fix MenuCommand_for_shifted_zoom() for 50%, which will have to work now.
(Previously, it was only used to disable the items for big zoom levels.)
Pass the raster_desc to command_handler() so we can use it when zooming.