v68k-screen:  Don't clear screen on exit
This was important when a front end crashed on exit and a pause occurred
while the crashlog was generated -- it was better to have the screen get
cleared immediately than to have the stale desktop image lingering.  But
now that the crash error has been fixed and we have a nice fadeout, it's
no longer desirable.  Furthermore, apparently the front end doesn't even
get notified about the update, and the cleared screen is only visible if
another update has occurred immediately before (e.g. to close a window),
which tends to result in tearing: partly empty desktop and partly black.